# apps/docs/apis/openapi.yaml
openapi: 3.1.0
info:
  title: AI GNT API
  version: 0.1.0
paths:
...
  /v1/settings/github:
    get:
      tags:
      - github
      summary: Get Github Connection
      operationId: get_github_connection_v1_settings_github_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
    post:
      tags:
      - github
      summary: Connect Github
      operationId: connect_github_v1_settings_github_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectGithubRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
    delete:
      tags:
      - github
      summary: Disconnect Github
      operationId: disconnect_github_v1_settings_github_delete
      responses:
        '204':
          description: Successful Response
      security:
      - HTTPBearer: []
  /v1/settings/github/app/install-url:
    get:
      tags:
      - github
      summary: Github App Install Url
      description: 'Starts the GitHub App install flow — mints a signed state token

        binding this install to org.org_id (see app_auth.verify_install_state

        for why the callback trusts this instead of whatever org happens to be

        active in the browser that completes it) and points at the App''s own

        install URL, fetched live rather than a hardcoded slug.'
      operationId: github_app_install_url_v1_settings_github_app_install_url_get
      security:
      - HTTPBearer: []
      parameters:
      - name: origin
        in: query
        required: false
        schema:
          type: string
          default: web
          title: Origin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/settings/github/app/callback:
    get:
      tags:
      - github
      summary: Github App Callback
      description: 'Called server-to-server by apps/web''s `/api/github/callback`
        route

        (the URL actually registered in the App''s settings — GitHub redirects

        the browser there, not here) once GitHub redirects back with a real

        installation_id. Every genuine install this app itself started carries

        `state` (see install-url above) — that signature IS the auth for this

        endpoint, there''s no session/API-key check here at all, same trust

        model as gnt/slack/oauth.py''s own oauth_callback. `setup_action` values

        other than "install" (a pending "request" awaiting an org owner''s

        approval, or a repo-selection "update" from GitHub''s own installation

        settings page, which doesn''t carry state at all) aren''t handled here —

        out of scope for this connect flow; re-running `gnt connect github`

        covers the update case in practice.'
      operationId: github_app_callback_v1_settings_github_app_callback_get
      parameters:
      - name: installation_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Installation Id
      - name: setup_action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Setup Action
      - name: state
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: State
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ConnectGithubRequest:
      properties:
        repo_url:
          type: string
          title: Repo Url
        pat:
          type: string
          format: password
          title: Pat
          writeOnly: true
      type: object
      required:
      - repo_url
      - pat
      title: ConnectGithubRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://api.gntai.dev
  description: Production